+++ processor/src/main/kotlin/io/mcarle/konvert/processor/codegen/MappingCodeGenerator.kt
@@ -1,12 +1,7 @@
 package io.mcarle.konvert.processor.codegen
 
-import com.google.devtools.ksp.symbol.KSClassDeclaration
-import com.google.devtools.ksp.symbol.KSFunctionDeclaration
-import com.google.devtools.ksp.symbol.KSPropertyDeclaration
-import com.google.devtools.ksp.symbol.KSType
-import com.google.devtools.ksp.symbol.KSTypeReference
-import com.google.devtools.ksp.symbol.KSValueParameter
-import com.google.devtools.ksp.symbol.Origin
+import com.google.devtools.ksp.processing.KSPLogger
+import com.google.devtools.ksp.symbol.*
 import com.squareup.kotlinpoet.CodeBlock
 import com.squareup.kotlinpoet.joinToCode
 import com.squareup.kotlinpoet.ksp.toClassName
@@ -14,14 +9,17 @@ import io.mcarle.konvert.converter.api.TypeConverterRegistry
 import io.mcarle.konvert.converter.api.config.Configuration
 import io.mcarle.konvert.converter.api.config.enableConverters
 import io.mcarle.konvert.converter.api.config.enforceNotNull
+import io.mcarle.konvert.converter.api.config.ignoreUnmappedTargetProperties
 import io.mcarle.konvert.converter.api.isNullable
 import io.mcarle.konvert.processor.exceptions.IgnoredTargetNotIgnorableException
 import io.mcarle.konvert.processor.exceptions.NoMatchingTypeConverterException
 import io.mcarle.konvert.processor.exceptions.NotNullOperatorNotEnabledException
 import io.mcarle.konvert.processor.exceptions.PropertyMappingNotExistingException
-import java.util.Locale
+import java.util.*
 
-class MappingCodeGenerator {
+class MappingCodeGenerator constructor(
+    private val logger: KSPLogger
+) {
 
     fun generateMappingCode(
         source: KSType,
@@ -164,16 +162,28 @@ $className(${"⇥\n%L"}
     ): CodeBlock {
         return targetProperties.mapNotNull { targetProperty ->
             val sourceProperty = determinePropertyMappingInfo(sourceProperties, targetProperty)
-            val convertedValue = convertValue(
-                source = sourceProperty,
-                targetTypeRef = targetProperty.type,
-                valueParamIsNullable = false,
-                valueParamHasDefault = true
-            )
-            if (convertedValue != null) {
-                CodeBlock.of("$targetVarName.${sourceProperty.targetName}·=·$convertedValue")
+            if (sourceProperty == null) {
+                if (!Configuration.ignoreUnmappedTargetProperties) {
+                    throw PropertyMappingNotExistingException(targetProperty, sourceProperties)
+                } else {
+                    logger.warn(
+                        "Ignoring unmapped target property `${targetProperty.simpleName.asString()}`  due to configuration 'konvert.ignore-unmapped-target-properties=true'",
+                        targetProperty
+                    )
+                    null
+                }
             } else {
-                null
+                val convertedValue = convertValue(
+                    source = sourceProperty,
+                    targetTypeRef = targetProperty.type,
+                    valueParamIsNullable = false,
+                    valueParamHasDefault = true
+                )
+                if (convertedValue != null) {
+                    CodeBlock.of("$targetVarName.${sourceProperty.targetName}·=·$convertedValue")
+                } else {
+                    null
+                }
             }
         }.joinToCode("\n")
     }
@@ -190,10 +200,10 @@ $className(${"⇥\n%L"}
     private fun determinePropertyMappingInfo(
         propertyMappings: List<PropertyMappingInfo>,
         ksPropertyDeclaration: KSPropertyDeclaration
-    ): PropertyMappingInfo {
+    ): PropertyMappingInfo? {
         return propertyMappings.firstOrNull {
             it.targetName == ksPropertyDeclaration.simpleName.asString()
-        } ?: throw PropertyMappingNotExistingException(ksPropertyDeclaration, propertyMappings)
+        }
     }
 
     private fun convertValue(